Define concrete parameters
I misunderstood how MOQT's parameers worked when first drafting things up.

This change splits the parameter space into the two types defined by MOQT:
"setup parameters" and "paramters". It then defines the concrete parameters
that have been defined in the MOQT draft so that its much easier dealing
with seralization and parsing, rather than having a generic idea of a
"parameter has bytes". This is channeling some of the work on
https://github.com/moq-wg/moq-transport/pull/761/files but doesn't go all
the way. I think this design is compatible with all approaches in MOQT
present or future.