[updates for draft-14] Modify subgroup_header_created and subgroup_header_parsed (#52)
* [updates for draft-14] Modify subgroup_header_created and subgroup_header_parsed

I've added the following two fields to the aforementioned events:
* contains_end_of_group
* extensions_present

Note that we don't log the wire-level "type" field of the subgroup header, but rather the semantic reconstructions of the information present in them. This is to ensure ease of debugging. Note that the subgroup_id field is null if and only if the "type" field of the subgroup header indicates that the subgroup id will be the id of the first object sent over that subgroup.

* Addressed comments 1

* Addressed comment 2

Co-authored-by: Lucas Pardue <lucas@lucaspardue.com>

* Addressed comment 3

Co-authored-by: Lucas Pardue <lucas@lucaspardue.com>

* Remove trailing whitespace

---------

Co-authored-by: Lucas Pardue <lucas@lucaspardue.com>