[updates for draft-14] Update structure of MOQTFetch (#56)
* [updates for draft-14] Update structure of MOQTFetch

In draft-14, MoQ reorganized FETCH. The message now carries either a "Standalone Fetch" or a "Joining Fetch" block. I made the structure in mLog more closely mirror that of MoQT, in order to fix a couple of issues (e.g. the fact that the track namespace wasn't optional earlier on).

* Make MOQTFetchType more specific (absolute vs relative join)

* Apply suggestions from code review (1)

Co-authored-by: Mathis Engelbart <mathis.engelbart@gmail.com>

---------

Co-authored-by: Mathis Engelbart <mathis.engelbart@gmail.com>